+++ offchain/crates/contributor-rewards/src/ingestor/mod.rs
@@ -4,5 +4,6 @@ pub mod fetcher;
 pub mod inet_accumulator;
 pub mod internet;
 pub mod serviceability;
+pub mod shred_subscription;
 pub mod telemetry;
 pub mod types;